requirements.txt
pillow==9.0.1
numpy==1.23.5
tqdm==4.64.0
opencv-python==4.6.0.66
taco-box==0.1.1
pandas==1.4.2
torch==2.0.0
torchvision
transformers==4.24.0
evaluate==0.2.2
